+++ src/pyiem/nws/products/saw.py
@@ -69,7 +69,7 @@ class SAWProduct(TextProduct):
         """Do the necessary database work
 
         Args:
-          (psycopg2.transaction): a database transaction
+          (psycopg.transaction): a database transaction
         """
         if self.action == self.ISSUES:
             # Ensure we have a watch to update